requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
yoomoney==0.18.0
pyyaml==6.0.1
aiohttp==3.9.1
colorlog==6.7.0
pytest==7.4.0